requirements.txt
asttokens==3.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
datadog-api-client==2.46.0
decorator==5.2.1
executing==2.2.1
fastjsonschema==2.21.2
git-filter-repo==2.47.0
idna==3.11
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
narwhals==2.13.0
nbformat==5.10.4
numpy==2.3.5
packaging==25.0
parso==0.8.5
platformdirs==4.5.1
plotly==6.5.0
prompt_toolkit==3.0.52
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
